Modified crepes.html
Added an image of a stack of crepes which is stored locally on the repository. The alt attribute is set to describe the picture accurately.

Added a description of the crepes recipe which include possible fillings (used unordered list) and prerequisites such as equipment, time expectations, etc.

Added all the ingredients needed for the crepes recipe using an unordered list.

Also added all the steps required for cooking the crepes using an ordered list.